# Conflicts:
#	mongo_mapper/document.py

+++ mongo_mapper/core/id_manager.py
@@ -2,6 +2,7 @@ from enum import Enum
 import uuid
 from pymongo import ReturnDocument
 from bson.objectid import ObjectId
+from bson.int64 import Int64
 from mongo_mapper.core.connection import get_collection
 
 
@@ -12,7 +13,7 @@ class IdType(Enum):
 def get_id(id_type, obj_name=None):
     if id_type is IdType.Numeric:
         _id = uuid.uuid1()
-        return _id.int >> 64
+        return Int64(_id.int >> 64)
     elif id_type is IdType.Incremental:
 
         collection = get_collection('', 'counters', True)
